Implement func approx [see description]
Utilized chain of responsibility pattern to handle
different types of the objects.
Handlers would be taken from the chain and
checked whether they can carry out the comparison.
If not, the next handler is used.

When dealing with mapping, sequence, set or composite objects
containing __dict__, the recursive invocation
of func approx() is used.